+++ src/bitbucket/bitbucket-server/pullRequests.ts
@@ -812,6 +812,7 @@ export class ServerPullRequestApi implements PullRequestApi {
                         name: accountId,
                     },
                 })),
+                draft: createPrData.draft || false,
             },
             {
                 markup: true,
@@ -839,6 +840,24 @@ export class ServerPullRequestApi implements PullRequestApi {
         const { data } = await this.client.put(
             `/rest/api/1.0/projects/${ownerSlug}/repos/${repoSlug}/pull-requests/${pr.data.id}`,
             prBody,
+            {
+                markup: true,
+                avatarSize: 64,
+            },
+        );
+
+        return ServerPullRequestApi.toPullRequestModel(data, 0, pr.site, pr.workspaceRepo);
+    }
+
+    async updateDraftStatus(pr: PullRequest, isDraft: boolean): Promise<PullRequest> {
+        const { ownerSlug, repoSlug } = pr.site;
+
+        const { data } = await this.client.put(
+            `/rest/api/1.0/projects/${ownerSlug}/repos/${repoSlug}/pull-requests/${pr.data.id}`,
+            {
+                version: pr.data.version,
+                draft: isDraft,
+            },
             {
                 markup: true,
                 avatarSize: 64,