+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ### [Report an Issue](https://github.com/atlassian/atlascode/issues)
 
 ## What's new in 4.0.20
+### Features
+
+- Added Draft pull requests
 
 ### Improvements